Move ActionText backend to own namespace
